CarRental (Back)/CarRental/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0
